Got the game working again
Organized and condensed required files.
Removed an unused resource.
Temporarily added names for the tanks.
Changed to having two tick handlers (one for the menu and one for the
game).
The Tank type no longer "fires" a missile, but instead it returns a
missile that can be put on stage.
Other various UI enhancements.
Code cleanup.